miniArca/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dnspython==2.7.0
exceptiongroup==1.3.0
fastapi==0.115.12
filelock==3.13.1
fonttools==4.58.1
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.72.1
grpcio-status==1.71.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.32.4
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.4
joblib==1.5.1
kiwisolver==1.4.8
MarkupSafe==2.1.5
matplotlib==3.10.3
motor==3.7.1
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.0
pillow==11.0.0
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
pymongo==4.13.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
sentence-transformers==4.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.2
sympy==1.13.1
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchvision==0.21.0+cu126
tqdm==4.67.1
transformers==4.52.4
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.150
ultralytics-thop==2.0.14
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.3
